Call parseTopLevelType on component command
