+++ packages/react-native-codegen/src/parsers/typescript/components/commands.js
@@ -15,19 +15,15 @@ import type {
   CommandTypeAnnotation,
 } from '../../../CodegenSchema.js';
 import type {TypeDeclarationMap} from '../utils.js';
-
-const {getValueFromTypes} = require('../utils.js');
+const {parseTopLevelType} = require('../parseTopLevelType');
 
 type EventTypeAST = Object;
 
 function buildCommandSchema(property: EventTypeAST, types: TypeDeclarationMap) {
+  const topLevelType = parseTopLevelType(property.typeAnnotation.typeAnnotation, types);
   const name = property.key.name;
-  const optional = property.optional || false;
-  const value = getValueFromTypes(
-    property.typeAnnotation.typeAnnotation,
-    types,
-  );
-
+  const optional = property.optional || topLevelType.optional;
+  const value = topLevelType.type;
   const firstParam = value.parameters[0].typeAnnotation;
 
   if (
@@ -45,10 +41,10 @@ function buildCommandSchema(property: EventTypeAST, types: TypeDeclarationMap) {
 
   const params = value.parameters.slice(1).map(param => {
     const paramName = param.name;
-    const paramValue = getValueFromTypes(
+    const paramValue = parseTopLevelType(
       param.typeAnnotation.typeAnnotation,
       types,
-    );
+    ).type;
 
     const type =
       paramValue.type === 'TSTypeReference'
